--- code/ModelConversionApp/Models/ColumnTypes.cs
@@ -1,22 +1,0 @@
-﻿namespace ModelConversionApp.Models;
-
-internal enum ColumnType
-{
-    Long,
-    String,
-    Timestamp
-}
-
-internal static class ColumnTypeConverter
-{
-    internal static string ConvertColumnTypeToString(ColumnType type)
-    {
-        return type switch
-        {
-            ColumnType.Long => "long",
-            ColumnType.String => "string",
-            ColumnType.Timestamp => "timestamp",
-            _ => throw new NotSupportedException()
-        };
-    }
-}
