--- code/ModelConversionApp/Models/DatabaseProperties.cs
@@ -1,9 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-namespace ModelConversionApp.Models;
-
-internal class DatabaseProperties
-{
-    [JsonPropertyName("IsSyMSCDMDatabase")]
-    public bool IsSyMSCDMDatabase { get; set; } = true;
-}
