--- code/ModelConversionApp/Models/Namespace.cs
@@ -1,9 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-namespace ModelConversionApp.Models;
-
-internal class Namespace
-{
-    [JsonPropertyName("DatabaseName")]
-    public string DatabaseName { get; set; }
-}
