--- code/ModelConversionApp/Models/Relationship.cs
@@ -1,23 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-namespace ModelConversionApp.Models;
-
-internal class Relationship
-{
-    [JsonPropertyName("Name")]
-    public string Name { get; set; }
-    [JsonPropertyName("EntityType")]
-    public string EntityType { get; set; } = "RELATIONSHIP";
-    [JsonPropertyName("Origin")]
-    public Origin Origin { get; set; } = new Origin();
-    [JsonPropertyName("Namespace")]
-    public Namespace Namespace { get; set; }
-    [JsonPropertyName("FromTableName")]
-    public string FromTableName { get; set; }
-    [JsonPropertyName("ToTableName")]
-    public string ToTableName { get; set; }
-    [JsonPropertyName("RelationshipType")]
-    public int RelationshipType { get; set; } = 0;
-    [JsonPropertyName("ColumnRelationshipInformations")]
-    public ColumnRelationshipInformations ColumnRelationshipInformations { get; set; }
-}
